@@ -1,26 +1,13 @@
 package io.horizen.evm;
 
 import io.horizen.evm.params.EvmParams;
-import io.horizen.evm.results.EvmResult;
-
-import java.math.BigInteger;
+import io.horizen.evm.results.InvocationResult;
 
 public final class Evm {
     private Evm() { }
 
-    public static EvmResult Apply(
-        ResourceHandle stateDBHandle,
-        Address from,
-        Address to,
-        BigInteger value,
-        byte[] input,
-        BigInteger gasLimit,
-        BigInteger gasPrice,
-        EvmContext context,
-        TraceOptions traceOptions
-    ) {
-        var params = new EvmParams(
-            stateDBHandle.handle, from, to, value, input, gasLimit, gasPrice, context, traceOptions);
-        return LibEvm.invoke("EvmApply", params, EvmResult.class);
+    public static InvocationResult Apply(ResourceHandle stateDBHandle, Invocation invocation, EvmContext context) {
+        var params = new EvmParams(stateDBHandle.handle, invocation, context);
+        return LibEvm.invoke("EvmApply", params, InvocationResult.class);
     }
 }
