@@ -1,40 +1,15 @@
 package io.horizen.evm.params;
 
-import io.horizen.evm.Address;
 import io.horizen.evm.EvmContext;
-import io.horizen.evm.TraceOptions;
-
-import java.math.BigInteger;
+import io.horizen.evm.Invocation;
 
 public class EvmParams extends HandleParams {
-    public final Address from;
-    public final Address to;
-    public final BigInteger value;
-    public final byte[] input;
-    public final BigInteger availableGas; // uint64
-    public final BigInteger gasPrice;
+    public final Invocation invocation;
     public final EvmContext context;
-    public final TraceOptions traceOptions;
 
-    public EvmParams(
-        int handle,
-        Address from,
-        Address to,
-        BigInteger value,
-        byte[] input,
-        BigInteger availableGas,
-        BigInteger gasPrice,
-        EvmContext context,
-        TraceOptions traceOptions
-    ) {
+    public EvmParams(int handle, Invocation invocation, EvmContext context) {
         super(handle);
-        this.from = from;
-        this.to = to;
-        this.value = value;
-        this.input = input;
-        this.availableGas = availableGas;
-        this.gasPrice = gasPrice;
+        this.invocation = invocation;
         this.context = context;
-        this.traceOptions = traceOptions;
     }
 }
