@@ -1,11 +1,11 @@
 package lib
 
 import (
-	"github.com/ethereum/go-ethereum/core/rawdb"
-	"github.com/ethereum/go-ethereum/core/state"
-	"github.com/ethereum/go-ethereum/ethdb"
-	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/trie"
+	"github.com/HorizenOfficial/go-ethereum/core/rawdb"
+	"github.com/HorizenOfficial/go-ethereum/core/state"
+	"github.com/HorizenOfficial/go-ethereum/ethdb"
+	"github.com/HorizenOfficial/go-ethereum/log"
+	"github.com/HorizenOfficial/go-ethereum/trie"
 )
 
 type Database struct {
@@ -29,12 +29,12 @@ func (s *Service) open(storage ethdb.Database) int {
 	return s.databases.Add(db)
 }
 
-func (s *Service) OpenMemoryDB() int {
+func (s *Service) DatabaseOpenMemoryDB() int {
 	log.Info("initializing memorydb")
 	return s.open(rawdb.NewMemoryDatabase())
 }
 
-func (s *Service) OpenLevelDB(params LevelDBParams) (error, int) {
+func (s *Service) DatabaseOpenLevelDB(params LevelDBParams) (error, int) {
 	log.Info("initializing leveldb", "path", params.Path)
 	storage, err := rawdb.NewLevelDBDatabase(params.Path, 256, 0, "zen/db/data/", false)
 	if err != nil {
@@ -44,7 +44,7 @@ func (s *Service) OpenLevelDB(params LevelDBParams) (error, int) {
 	return nil, s.open(storage)
 }
 
-func (s *Service) CloseDatabase(params DatabaseParams) error {
+func (s *Service) DatabaseClose(params DatabaseParams) error {
 	err, db := s.databases.Get(params.DatabaseHandle)
 	if err != nil {
 		return err
