@@ -1,9 +1,9 @@
 package lib
 
 import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/ethereum/go-ethereum/trie"
+	"github.com/HorizenOfficial/go-ethereum/common"
+	"github.com/HorizenOfficial/go-ethereum/rlp"
+	"github.com/HorizenOfficial/go-ethereum/trie"
 )
 
 type HashParams struct {
