@@ -1,13 +1,13 @@
 package lib
 
 import (
-	"github.com/ethereum/go-ethereum/common"
+	"github.com/HorizenOfficial/go-ethereum/common"
 )
 
 func SetupTest() (*Service, int, int) {
 	var (
 		instance       = New()
-		dbHandle       = instance.OpenMemoryDB()
+		dbHandle       = instance.DatabaseOpenMemoryDB()
 		_, stateHandle = instance.StateOpen(StateParams{
 			DatabaseParams: DatabaseParams{DatabaseHandle: dbHandle},
 			Root:           common.Hash{},
